Fix: 32-bit Model Index support GSTL Array format (layered texture) PCK v3 format

+++ src/process/scene/mesh/mesh_get_arrays.ts
@@ -445,7 +445,7 @@ export function _get_array_from_surface(p_format: number, p_vertex_data: ArrayBu
             }
           } else {
             const w = arr;
-            const ir = new Uint16Array(p_index_data);
+            const ir = new Uint32Array(p_index_data);
 
             for (let j = 0; j < p_index_len; j++) {
               const v = ir[j];
